Also move FailedUnicode to powerline/lib/unicode
